Separate ignition.go into more specific files
Previously ignition.go contained code for general ignition parsing, code
for generating install manifests using opesnshift-install, and code for
creating the discovery ignition.

The discovery ignition and installer manifests code has mostly nothing
to do with each other so those should be separated out into their own
files.

ignition.go remains as a place for ignition parsing code as well as the
very small number of common functions between the installer manifests
and discovery ignition use cases.
